refactor: move some code
