+++ OUDS/Core/Components/Sources/_/Resources/Localizable.xcstrings
@@ -5,7 +5,7 @@
       "shouldTranslate" : false
     },
     "%lld%%" : {
-      "comment" : "A percentage value describing the progress of a task.",
+      "comment" : "A progress indicator that shows the percentage of completion.",
       "isCommentAutoGenerated" : true
     },
     "core_alertMessage_close_a11y" : {
